Move CaseDiffTool to its own module
